requirements.txt
line-bot-sdk==3.8.0
openai==1.12.0
gspread==5.12.4
oauth2client==4.1.3
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0